Changed the type of subject credit points from int to double to accommodate half points. This is retroactive. -> Reflected in the sql files and subject lsf model. Renamed searchtools scss to bars to reflect its use in the tool bar the filter bar and the title bar. Included bars scss in item scss to suppress icon output. Moved standard (array) attribute output from subject item view to item view. Expanded list type attributes to allow for the linking of resources as appropriate. Placed the language selection form above the component title output. Expanded the action to include the component. This segment is automatically removed if a menu item of the component can be resolved. Changed the permissions check for item views from allow edit to allow view to better coincide with the user's experience.

+++ com_thm_organizer/site/Models/SubjectLSF.php
@@ -375,7 +375,7 @@ class SubjectLSF extends BaseModel
             return;
         }
 
-        $crp = (int)$dataObject->lp;
+        $crp = (float)$dataObject->lp;
 
         $this->setAttribute($subject, 'creditpoints', $crp);
         $this->crp = $crp;
